task: Add recursive SideNav
I've added a pair of componets that use a recursive renderer for the
tree data. This will take any level of nesting.
Added a bit of state at the top SideNavRecursive level called
`handleItemClick`. This will allow you to send messages back to the
server or do whatever else you want to do to the state in that
component.